build

+++ app/Http/Controllers/Favorites/FavoriteController.php
@@ -3,6 +3,7 @@
 namespace Azizner\Http\Controllers\Favorites;
 
 use Azizner\Announcement;
+use Azizner\Photo;
 use Illuminate\Http\Request;
 use Azizner\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -24,17 +25,32 @@ class FavoriteController extends Controller
     protected function validator(array $data)
     {
         $rules = [
-            'id' => ['nullable', 'numeric'],
+            'id' => ['required', 'numeric'],
+            'type' => ['required', 'string', 'in:announcement,photo'],
         ];
 
         return Validator::make($data, $rules);
     }
 
-    public function index()
+
+
+    public function announcements()
     {
         $user = Auth::user();
         $announcements = $user->favoriteAnnouncements()->paginate(2);
-        return view('favorites.index', ['announcements'=>$announcements]);
+        return view('favorites.announcements', ['announcements'=>$announcements]);
+    }
+
+    public function photos()
+    {
+        $user = Auth::user();
+        $photos = $user->favoritePhotos()->paginate(2);
+        return view('favorites.photos', ['photos'=>$photos]);
+    }
+
+    public function index()
+    {
+
     }
 
     /**
@@ -59,29 +75,59 @@ class FavoriteController extends Controller
 
         $validator = $this->validator($request->all());
         if($validator->fails()){
-            $response["status"] = "error";
+            $response["status"] = 'error';
             echo json_encode($response);
             exit;
         }
 
         $user = Auth::user();
 
-        if(!empty($announcement = Announcement::where('id',$request->get('id'))->first()))
-        {
-            if($user->favoriteAnnouncements->contains($announcement->id))
+        if($request->get('type') === 'announcement'){
+            if(!empty($announcement = Announcement::where('id',$request->get('id'))->first()))
             {
-                $user->favoriteAnnouncements()->detach($announcement->id);
-            }else{
-                $user->favoriteAnnouncements()->attach($announcement->id);
-            }
-            $response["status"] = "success";
+                if($user->favoriteAnnouncements->contains($announcement->id))
+                {
+                    $user->favoriteAnnouncements()->detach($announcement->id);
+                }else{
+                    $user->favoriteAnnouncements()->attach($announcement->id);
+                }
+                $response["status"] = "success";
 
-        }else{
-            $response["status"] = "error";
+            }else{
+                $response["status"] = "error";
+            }
+
+            echo json_encode($response);
+            exit;
         }
 
+        if($request->get('type') === 'photo') {
+            if(!empty($photo = Photo::where('id',$request->get('id'))->first()))
+            {
+                if($user->favoritePhotos->contains($photo->id))
+                {
+                    $user->favoritePhotos()->detach($photo->id);
+                }else{
+                    $user->favoritePhotos()->attach($photo->id);
+                }
+                $response["status"] = "success";
+
+            }else{
+                $response["status"] = "error";
+            }
+
+            echo json_encode($response);
+            exit;
+        }
+
+
+        $response["status"] = 'error';
         echo json_encode($response);
         exit;
+
+
+
+
     }
 
     /**
